+++ src/concepts/interfaces/get-concept-by-code-options.ts
@@ -1,0 +1,7 @@
+import type { VocabReleaseMixin } from '../../common/interfaces/vocab-release.js';
+
+export interface GetConceptByCodeOptions extends VocabReleaseMixin {
+  includeRelationships?: boolean;
+  includeSynonyms?: boolean;
+  includeHierarchy?: boolean;
+}